gui/web/src/MouseTrapBackendFiles/returnPathPredictions/requirements.txt
# Function dependencies, for example:
# package>=version

click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pip==18.0
requests==2.21.0
setuptools==40.2.0
Werkzeug==0.14.1
wheel==0.31.1
gcloud
google-cloud-bigquery
google-cloud-storage
google-api-python-client